services/voice-gemini/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-socketio>=5.10.0
python-multipart>=0.0.6

# Google GenAI SDK
google-genai>=0.3.0

# Audio
pyaudio>=0.2.14
pillow>=10.0.0

# Environment
python-dotenv>=1.0.0

# Async HTTP
httpx>=0.25.0
aiohttp>=3.9.0
